backend/requirements.txt
fastapi
uvicorn
youtube-transcript-api
langchain
langchain-community
langchain-text-splitters
langchain-google-genai
faiss-cpu
python-dotenv